Code Cleanup

+++ GameProject/GameProject/Game/Sounds.cs
@@ -1,11 +1,6 @@
 ﻿using Microsoft.Xna.Framework.Audio;
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Media;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace GameProject
 {
@@ -25,7 +20,7 @@ namespace GameProject
         }             
      
         //DRAW
-        public static void playBackgroundMusic(int volume)
+        public static void playBackgroundMusic(float volume)
         {
             MediaPlayer.Volume = volume;
             MediaPlayer.Play(BackgroundMusic);
